+++ adventofcode2019/src/nl/michielgraat/adventofcode2019/intcode/instructions/JumpIfTrue.java
@@ -6,7 +6,7 @@ public class JumpIfTrue extends Instruction {
 
     int ptrIncr;
 
-    public JumpIfTrue(final List<Integer> memory, final int ptr, final int modes) {
+    public JumpIfTrue(final List<Long> memory, final int ptr, final int modes) {
         super(memory, ptr, modes);
         this.ptrIncr = 3;
     }
@@ -27,7 +27,7 @@ public class JumpIfTrue extends Instruction {
     }
 
     @Override
-    public List<Integer> getInputPositions() {
+    public List<Long> getInputPositions() {
         return memory.subList(ptr + 1, ptr + 3);
     }
 
@@ -43,19 +43,19 @@ public class JumpIfTrue extends Instruction {
 
     @Override
     public void execute() {
-        int var1 = 0;
-        int var2 = 0;
+        long var1 = 0;
+        long var2 = 0;
         if (modes > -1) {
             final int mode1 = modes % 10;
             final int mode2 = (modes / 10) % 10;
-            var1 = mode1 == 0 ? memory.get(memory.get(ptr + 1)) : memory.get(ptr + 1);
-            var2 = mode2 == 0 ? memory.get(memory.get(ptr + 2)) : memory.get(ptr + 2);
+            var1 = mode1 == 0 ? memory.get(memory.get(ptr + 1).intValue()) : memory.get(ptr + 1);
+            var2 = mode2 == 0 ? memory.get(memory.get(ptr + 2).intValue()) : memory.get(ptr + 2);
         } else {
-            var1 = memory.get(memory.get(ptr + 1));
-            var2 = memory.get(memory.get(ptr + 2));
+            var1 = memory.get(memory.get(ptr + 1).intValue());
+            var2 = memory.get(memory.get(ptr + 2).intValue());
         }
         if (var1 != 0) {
-            ptrIncr = var2 - ptr;
+            ptrIncr = (int) var2 - ptr;
         }
     }
